backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
sqlalchemy[asyncio]==2.0.41
asyncpg==0.31.0
alembic==1.15.2
yahooquery>=2.3.7
pandas==2.2.3
apscheduler==3.11.0
pydantic-settings==2.9.1
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
aiosqlite==0.21.0
